Fix race in test_lost_merge_requests
At the end of this test, we check the merger api cache to see if the
merge requset is cleaned up.  We may need to give the cache time to
update, so use iterate_timeout.

We should also be consistent about which merger api object we use;
to that end, the test is updated to only use the api object owned by
the merger client when checking the cache contents (since that's the
one that we use to run the cleanup method).

Finally, it appears this was subject to the same issue that was fixed
in 4f96125007ebfe4e8ccd634565313114009bc46d where we were not handling
deleted znodes correctly.

Change-Id: Ic39b01b2d2eb138cb79c7f5b412c6390c07112bf