Join the scheduler thread before stopping the zk client
The change to add zk locking inside the main scheduler processing
loop conflicts with the shutdown procedure.  Currently the following
sequence happens:

[1] scheduler.stop() called
[1] internal stop flag is set
[1] scheduler wake event is set
[2] scheduler main thread wakes up
[2] scheduler main thread obtains zk lock
[1] remainder of scheduler.stop() method runs, stopping other threads
[1] scheduler.stop calls self.zk_client.disconnect()
[2] scheduler main thread finishes main loop
[2] scheduler main thread releases zk lock, which fails due to no conn

This is mostly harmless, however it does spew lots of tracebacks in
the unit test logs.  To avoid this, join the main thread before stopping
the zk connection.

Change-Id: I0f7c96d18ea7ea269d453e67f7c6da1e7be9bea3