Update patch set 3
Patch Set 3:

Would it work if we include some images (basically what [1]says)?

We could - for example - provide some icons via zuul-web, look up those URLs in the github reporter and put them in the markdown description for the check run. But maybe that's too complex and too much effort to work around the limitation.

I also did some investigation and Travis CI seems to be using a different field to provide the check run details. They are also including some collapsable elements (Build Configuration [2]) in their description, which IMHO cannot be done with plain markdown. You can check for that if you inspect the details div of a Zuul check run [3] and the details of a Travis CI check run [2] in the browser dev tools. Zuul is putting everything into <div class="markdown-body" ...> and Travis CI is putting there only the title, but everything else is put into another <div> on the same level.

Because of that I checked out the github checks API documentation and it looks like the check run output also has another field "text" next to "summary" [4] (but the documentation says for both "This parameter supports Markdown", so I'd assume they come with the same limitations - but maybe we can give that a try just to make sure).

[1]: https://github.com/github/markup/issues/1440#issuecomment-803889380
[2]: https://github.com/pypa/pip/pull/9107/checks?check_run_id=1387918236
[3]: https://github.com/pypa/pip/pull/9107/checks?check_run_id=1387918396
[4]: https://docs.github.com/en/rest/reference/checks#output-object

Patch-set: 3