Fix cleanup thread not starting
The recent change to avoid reconfiguration on startup also meant
that we didn't release the hold on the cleanup thread (we need to
have a working configuration in order to know, for example, what
semaphores to clean up).  To correct this, use the timestamp of
the config priming as the last "reconfiguration" time.

Change-Id: I1c96618524c6a32dbc688319a82db11ddc94210b