Suppress spurious semaphore lock error logs
When a job is unable to acquire all the semaphores it needs, it
unlocks all of its semaphores and waits until the next attempt.

It does this by performing a naive release of all of the semaphores
it knows about.  If it has them, they will be released.  If it does
not, then it will be a no-op.  But we log the "no-op" case as an
error.  Add an extra argument to indicate that we expect that to
fail and we don't want to log a release failure as an error in
that case.

Change-Id: If204c7cdcad009155fe1fbe30b34d1bdd14a2703