Use branchesForRepoState in all cases
This method returns a reduced set of branches for Zuul to use when
preparing the repo state for running jobs.  If exclude-unprotected-
branches was set, it would return the superset of all possible
branches in all the involved repos.  If unset, it would return None
meaning it should be unfiltered.

To allow include-branches to have the same effect, remove the check
on exclude-unprotected-branches.  This means that the method will
always be used, and the set of branches will always be sent to the
merger.  This may be a little more verbose, but it's simple and
should produce the same result.

Change-Id: I70e7c55203294b6c8eb852c623b6108967c2d322