Update patch set 7
Patch Set 7: Code-Review-1

We should keep colorblindness in mind, especially red-green colorblindness.

There are a few general principles to help with this:
* Usually adding some blue to the green color can help mitigate this (but if blue is used for other purposes, that can be less effective).
* Don't rely only on the color, but also use shapes or words.

We don't control the color of the green icons, so even if they do have some blue (I haven't checked), I don't think we should rely entirely on that.

We still have the words "success/failure" but the whole point of this exercise is to make it easier to scan and draw attention to failures; we should try to benefit colorblind people as well.

For those reasons, I think we should go with the check/X marks instead of the colored balls.

But instead of the white check, why don't we use the green one (heavy_check_mark)?

I posted that as solution 13 on https://github.com/dozhang/test/pull/1

Incidentally, the weird " : " thing was added originally to help with gerrit formatting using commentlinks.  We can drop it (or otherwise alter it) in github formatting.

Also, doing whatever you did in solution 12 to remove the bullet points might be nice too.

Patch-set: 7
Reviewer: Gerrit User 1 <1@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1