web: use spread operator in JOB_FETCH_SUCCESS reducer
I am seeing

 A state mutation was detected inside a dispatch, in the path:
 `job.jobs`. Take a look at the reducer(s) handling the action
 {"type":"JOB_FETCH_SUCCESS" ...

pop up when trying to make some changes to this page.  I am guessing
the update() and $merge were not as immuitable as we hoped.  This may
have started happening with Firefox 90.0.2 but I'm not 100% sure.

Writing this with spread operators seems more common these days, and
makes the error go away for me.

Change-Id: I324222467eea26135d3ab013831c4522e96306e3