web: Job: remove height setting for job page
This has been in since the original change
Idbeae3a11ec4180a193923def7dc7f9c53dc9043 but I couldn't trigger
anything that seemed to mess up the height on the display page.

This removes it to make the following changes that refactor this page
to PF4 compontents a bit easier to follow.  These changes introduce
new tables components, etc. and I haven't seen any need for adusting
the height.

Change-Id: I5fcc8c9ebc1da8b37526fe2c4be630ef8f1945ad