web: JobVariant: use @patternfly/react-table
Update this to use @patternfly/react-table; apparently the new hotness
is TableComposable so use that.

There is no option in this table component for vertical dividers.  To
maintain a visual difference between heading and data, and for
consistency with the build page, we bold the description column.

This includes the results of a 'yarn upgrade @patternfly/react-table
--latest' run for good measure.

Change-Id: I2f28d49ba1dee6b8d487cf7c5b1043b42f649c9f