Fix ordering of parameters in _logQueueStatus() call
The parameters to _logQueueStatus() are provided in the wrong order.
This put me on a completely wrong track when debugging a test case in
Zuul that didn't settle in time as the queue status mentioned "All
builds waiting" to be the culprit and not "All requests completed".

The output of _logQueueStatus() was:

    All ZK event queues empty: True
    All merge jobs waiting: True
    All builds reported: True
    All builds waiting: False
    All requests completed: True
    All event queues empty: True

while it should have been:

    All ZK event queues empty: True
    All merge jobs waiting: True
    All builds reported: True
    All builds waiting: True
    All requests completed: False
    All event queues empty: True

Change-Id: I161b8bf2aee208033730a4e29606887a60eb5768