Update patch set 2
Patch Set 2:

I like what this does for jobs with variants as it makes it clearer there are selections to view. However we get an empty tab when there are no variants that looks like a little bookmark at the top. Can we render that differently to make it clear there is a single version of the job? To see this view the airship-deploy-sip job in the openstack client under the site preview.

Patch-set: 2
CC: Gerrit User 4146 <4146@4a232e18-c5a9-48ee-94c0-e04e7cca6543>