Update patch set 2
Patch Set 2:

If we have a table, then having vertical bars would be nice.

But as an alternative, let's consider removing the horizontal bars too!

We have a really nice table-like layout on the build page for similar information, and it has no lines.  If this page can handle that, it would be my preference.

Patch-set: 2
CC: Gerrit User 1 <1@4a232e18-c5a9-48ee-94c0-e04e7cca6543>