Update patch set 2
Patch Set 2:

> This change removes the bar between columns in the JobVariant table. Visually having that vertical separator is nice and we should add it back in.

AFAICT the table doesn't have any way to set borders on individual cells.

I think patternfly would probably say that this is an incorrect usage of a table; as the first column is really more of a header (description of the column next to it) and it should be a data list.  

There is a newish (I think) model called a "Description List" which would actually be better.  It fits the model (a bunch of things with a desription).  I can propose this as a follow-on and we can discuss if that fits better.

In the mean time, I can make the first column <strong>, that would be similar to the build page.


[1] https://review.opendev.org/c/zuul/zuul/+/804479

Patch-set: 2