Perform an implicit smart-reconfigure on startup
If the scheduler is stopped, and new projects are added to an existing
tenant's configuration, then the scheduler is started again, the updated
configuration will not take effect because the scheduler considers the
cached configuration valid.

This may be surprising to users, so perform an implicit smart-reconfiguration
after priming the scheduler to pick up any changes.  This should be a
quick no-op if there are no changes.

Note: the zuul-operator essentially does this during its functional test
due to the sequencing involved.

Change-Id: I125e1da0243efff85eea42d7677ecf58b5dafa31