Update patch set 1
Patch Set 1:

> I really like this, but i think we need to reduce the amount of space between rows.

I have done this and reduced it from 1.5rem to 

 style={{'--pf-c-description-list--RowGap': '0.5rem'}}

of course that could be fiddled with

> Also it would be good to reduce the leading (the amount of space between lines) in text bodies (like the description), but that's less important.  

I'm coming to the opinion we should pull the description out of this and show it separately as done in the follow-on.  Some things, like the base job, are very long.  They are pre-wrapped for display in code editors (and show with embedded RST) which doesn't display well.  Having the description as a separate component at least lays some groundwork for others to make this even better.

> Similarly, the vertical space between nodeset elements and roles should be smaller.

The treeview doesn't have a compact(er) layout.  I think it looks better in the column layout

> "Semaphores required" consistently wraps; maybe we should drop the word "required"?

This could be expanded, the default width seems to be 12ch.  less is probably more; dropping the "required" is good.

Patch-set: 1