web: JobVariant : use wide layout for DescriptionList
Use a wide-layout with columns which I think makes good use of
available space.  This worked quite well, except for the job
description which varies between nothing, one-line and quite a wall of
text.  Large descriptions can really throw out the column balancing
and make it look weird.  On balance I think the job description is one
of the more interesting parts, so I pulled it out into it's own card
at the top, which starts expanded but can be rolled-up if it's
particuarly long.  I definitely think you could do more interesting
things (left some notes) but it's a start.

Change-Id: I315d338d8ebbb991bb3771238de0f43c59348510