web: Jobs: Use TreeView for job overview page
This converts this page to use the TreeView element with inline
search.

I couldn't get the extant code to format things quite the way I needed
for the TreeView, and it didn't seem to record branches.  I think this
is substantially the same in it's operation however.  I've probably
over-commented a bit; but building a tree backwards is a bit of a
non-obvious operation.

Change-Id: I2a2a16337d11e431445dc669ea6a19b934d836a2