Fix error with config cache some more
The previous commit to fix errors with the commit cache overlooked
the fact that the merger returns explicit dictionary entries for
all filenames specifically requested.  So to determine that a file
under zuul.d/ does not exist, it is sufficient to check if the
path is in the files dict returned by the merger.  But it is
insufficient to perform the same check for zuul.yaml since it
is a specific file requested, not a dir.

The merger does return None as the content of those entries if
the file does not exist.  To accomodate this, test for a None value
in addition to membership.  Also, add a test to cover this case.

Additionally, it was noted that the new debug log entry was reporting
the wrong project name.  This did not affect the cache cleanup, but
the same value is used to lock the cache.  This is because both lines
were referencing the leftover "project" variable from the loop above.
Both lines are updated to use the correct variable, and the project
variable is removed from scope to avoid future confusion.

Change-Id: Ie723d13a3ec5156ff19fbb695c5314e46aee2e02