Fix config cache cleanup with extra config paths
A cat job returns all required config files for ALL tenants a project is
part of. Since the config cache cleanup was only considering extra
config paths from its own tenant project config (tpc) we could have
ended up with stale extra config files in the cache depending on which
tenant triggered the cat job and performed the cleanup.

We can simplify the cache cleanup and solve the described issue by
clearing the whole project-branch cache before populating it with the
updated content from the cat job.

Since the issue depends on the timing which tenant reconfiguration is
process first, the added test case might not always show the correct
sequence of tenant reconfigurations for reproducing this bug.

The order of tenant reconfigs that produced the issue was:
    1. "tenant-one": no extra config paths and with that no cleanup of
       those old entries
    2. "tenant-two": uses the cached result and after the next restart
       the outdated entries from the cache

Change-Id: I0f82f7e25958fb4fa5abeb800a791eb57636eb50