web: Nodes: use composable table
This converts the node overview page to use a composable table.

These tables support adding a sort-order, which required some
reworking.  From what I've read, I believe that the correct way to
incorporate sorting like this with Redux is to implement it in the
reducer, so that the display page is not mutating the state of the
node list returned from the API in any way.

To this end we have a new action NODES_SORT which we trigger when the
table sort event fires.  This keeps track of the column being sorted
and direction in the state, and reorders the node list as required.

The NODES_FETCH_SUCCESS reducer is updated to also obey the current
sort ordering.  The end result is that if you sort by a column, then
press the "refresh" button you should get fresh data with the sorting
maintained.

Change-Id: Ibcf260fda08486b489797ea8073e7a057d7040cd