web: Projects: use composable table
This converts the projects overview page to a composable table.

Sorting is enabled for relevant fields.  A filter search is also
added.

Similar to Ibcf260fda08486b489797ea8073e7a057d7040cd the sorting state
is kept in redux.  However the filter is done via a selector, which I
believe to be the correct pattern to use.

Change-Id: I4b073fbd1a0ee36483158a74fbc8c99ef9842c43