Add source interface for setting change attributes
Changes are mostly created and updated by the drivers. However, since
there are some change attributes that are also modified from other parts
of the code, we need to make sure to update the cache in Zookeeper in
those cases. For this we introduce `setChangeAttributes()` as an
additional `Source` interface.

Change-Id: Iab9bc4a6e40f254c1cbc4405e90cb5f03e3ecd56