Add source interface for setting change attributes
Changes are mostly created and updated by the drivers. However, since
there are some change attributes that are also modified from other parts
of the code, we need to make sure to update the cache in Zookeeper in
those cases. For this we introduce `setChangeAttributes()` as an
additional `Source` interface.

Attributes of non-`Change` refs can still be set directly as those are
immutable from a driver's perspective. They will be stored as part of
the `QueueItem` state in Zookeeper.

Change-Id: Iab9bc4a6e40f254c1cbc4405e90cb5f03e3ecd56