wip: Add source interfaces for cached changes
For working with cached changes in a multi-schedule setup we need a way
to get changes using an Zuul internal cache key. A change's cache key
consists of the connection name for a particular change and a opaque key
that is choosen by the driver.

Changes are mostly created and updated by the drivers. However, since
there are some change attributes that are also modified from other parts
of the code, we need to make sure to update the cache in those cases.
For this we need to introduce additional `Source` interfaces.

New interfaces are:
* getChangeByKey(): get a change by a Zuul internal change key
* setFiles(): allows updating change.files
* setDependencies(): allows updating change.commit_needs_changes
* setRefreshDeps(): allows updating change.refresh_deps

TODO: Clarify if we also need an interface for updating the
`containing_branches` attribute or if this should be stored on the queue
item instead. In case we store this on the queue item, we need to modify
the change matcher so they can access this attribute.

    zuul/manager/__init__.py
    1476:        item.change.containing_branches = event.item_in_branches

Change-Id: Iab9bc4a6e40f254c1cbc4405e90cb5f03e3ecd56