Improve gitlab request logging
Adjust some log levels between info/debug to more closely match
the way it's used elsewhere in Zuul.

Add some text to the event payload log line so that we're not
just emitting a row JSON blob with no information.

Generate the event uuid early so that we can use it in the gitlab
connector.  This makes it easy to trace events all the way back
to ingestion.

Change-Id: I578de130b41cf84e8a415ef81e4efc69c1941c22