web: Project: cleanup query
Similar to I324222467eea26135d3ab013831c4522e96306e3 this is giving me
immutabilty errors in Firefox.

Looking closer, I'm not sure what this is doing.  It appears to be
getting the project, then putting it back into a "projects" array
indexed via tenants ... I'm assuming this was code reuse from the
"projects" (note "s") version or something.

All it really needs to do is pass back the project object from the
API.  Simplify it to do this.

Change-Id: I987446d2ab36727cc717f6cb87f82c50a59ec5c4