wip: Resolve change dependency references lazily
The Change class is modified so that references to dependencies are
evaluated lazily in order to prevent infinite recursion in case of
circular references. In order to resolve those references the change
needs to have access to the connection registry.

A different approach would be to always reference other changes via the
cache key and resolve those externally, e.g. in the pipeline manager.

Change-Id: I09845d65864edc0e54af5f24d3c7be8fe2f7a919