Update patch set 1
Patch Set 1: Code-Review-1

This does what is expected. However it feels a bit "overkill": the sort order only really matters on the Nodes page, while the redux store shares the nodes list with the whole application. If someday this data is to be used on other parts of the GUI, this will incur complications as the list may be re-sorted in an unexpected way. But then again, the nodes list might never be used anywhere else.

Wouldn't it be possible to hold a copy of the nodes in the state of the NodesPage component instead, and handle the sorting directly in the component as shown in patternfly's documentation ? [1] 

[1] https://www.patternfly.org/v4/components/table#composable-sortable--wrapping-headers

Patch-set: 1
Reviewer: Gerrit User 7186 <7186@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1