Remove returnNodeSet calls from scheduler
Now that nodes are locked on the executor, there are no cases wher
returnNodeSet does anything useful on the scheduler.  The code in
returnNodeSet is already a noop in this case: it does nothing if
the Node objects don't have a lock (and the scheduler never holds
one), and it short-circuits the resource counting if an AnsibleJob
is not provided (and the scheduler does not provide one).

Change-Id: Icdb26eccebb7bd1e856532306c97842f740e69c1