Fix failed merge detection in gitlab
gitlab is funny in that it can return status_code=200 with an error in
the body. This can happen when MR accept request fails (i.e. when
project requires squashing, but we have not done this). Ensure status of
the MR is "merged" otherwise return "merge_error" attribute as a failure
description.

Change-Id: I343fb08cdf58fdf6d2068e274a89a34a55adde5a