Update bindep for debian bullseye
The container images are based on opendevorg/python-base which is
based on python which is based on debian. Upstream python image has
updated from buster to bullseye. Once that makes it through
opendevorg/python-base, our images will start being bullseye. Make
the bindep entries forward compatible with such a future.

Change-Id: I71182e9d3e6e930977a9f983b37743ee3300ec91