Make node requests persistent
The original Nodepool protocol specified that node requests should
be ephemeral, that way if the requestor crashed before accepting
the nodes, the request would automatically be cleaned up and the
nodes returned.  This doesn't comport with multiple schedulers, as
we will soon expect schedulers to stop and start routinely while
we want the node requests they spawn to persist and be handled by
other schedulers.

Fortunately, Nodepool doesn't really care if the request is
ephemeral or not.  So we'll drop the "ephemeral" flag.

But in the short term, we will be stopping the scheduler and that
will leave orphan node requests.  And even in the long term, we
may have a complete Zuul system shutdown or even a bug which may
leak node requests, so we still need a way of deleting node requests
which don't belong.  To handle that, we add a cleanup routine which
we run immediately on startup and every hour that looks for node
requests created by this Zuul system but don't correspond to any
queue entries.  We create a new UUID to identify the Zuul system
and store it in ZK (so that if Nodepool has any other users we
don't delete their requests).

We no longer need to resubmit requests on connection loss, so tests
addressing that behavior are removed.

Change-Id: Ie22e99ef71cbe6b31d40c25a21498c1e867ca777