Update patch set 1
Patch Set 1: Code-Review-1

I'm reading this in the context of I83d89c2903e6fd0aaf67c66f61d106b2c5db3797 and it seems like it was intentional to show these

I'm -1 because I feel like the intent of above to show the retries is good and we should be communicating clearly that there are retries happening.

Honestly the example you linked to (11e9de4b452c4670b0783b01bb2c0553 in opendev) is probably like worse-case and I'm not sure the retry attempts are overwhelming in the info there.  So there is some argument for just keeping as is...

As a suggestion, something that would be clear but hide the jobs by default -- I think you could probably keep a list of non-final jobs, and then for each final job we are showing, if it has entries in the non-final list add an alert notification (bell type thing) in a column.  when you click on the alert it can expand the retry jobs.

Patch-set: 1
Reviewer: Gerrit User 7118 <7118@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1