Be more liberal with Depends-On: parsing
I just spent quite a while debugging only to realise I'd written
"Depends-On:  https://" (note the two spaces...).

This gets passed to urlparse.parse.urlparse() which does not handle
leading spaces; thus the return is considered invalid and the
Depends-On gets silently ignored.

It seems we should just be liberal with leading and trailing space
around Depends-On: lines.  This strips the results to acheive this,
and adds a small unit-test to cover this.

Change-Id: Iddd8363580fb2fefd99d4783fc90f68f34f0c949