Periodically maintain connection caches
With the persistent change cache in Zookeeper we need to periodically
remove changes that are outdated and no longer required by any active
item in the pipelines.

We also need to remove dangling change data nodes that have been
identified in the last run. The reason for this delayed cleanup is that
we don't want to remove data nodes where a driver is in the process of
creating/updating a new cache entry.  Alternativelay we could also use a
time based approach here.

The cache maintenance will be performed together with the other general
cleanup every hour.

Change-Id: I62e75ab8c5b43f830e01b0e4c08b25ecdc5eed08