Wrap nodepool request completed events with election
So that only one scheduler puts nodepool request completed events
in the queue, wrap that with an election.  There is a dedicated
thread to try to win the election, and if it does, it emits
complete events for every completed request (in case we missed
some during the handover).  Other than that, the process stays
the same.

If we encounter a problem putting the event on the ZK queue, we
tell the election thread to re-run the election.

Change-Id: I3dadf5524dc3d931415e20267d36030e945a3000