Remove hold request cache from nodepool interface
The hold request cache is always running on every Zuul component
that connects to nodepool, but it's actually unused.  Since there
are typically a small number of hold requests, it seems like asking
ZK for the list each time we finish a job isn't a big cost, so
let's remove it.  It's easy enough to add in (and actually fully
implement) if we want to in the future.

However, we are about to add a request cache, so the framework
to support that is left in place with comments.

Change-Id: Ic341db5c195a15d62845181c6a759664f56b64ce