Don't store node requests/nodesets on queue items
To prepare for queue items moving into ZooKeeper, stop storing the
NodeRequest and NodeSet objects on them.  Instead, reference requests
by ID and consult ZK when necessary, and store only the info about
nodesets that the scheduler needs.  The result are simple dicts
than can easily be serialized.

The deleteNodeRequest method is updated to accept IDs instead of
NodeRequest objects to minimize the number of times we need to
use a full NodeRequest object.

Change-Id: I3587a42eb5a151f41369385e482b7f36b1c41bf6