Remove internal nodepool request cache
The internal zuul.nodepool.Nodepool.requests dictionary is used so
the scheduler can keep track of its requests.  Since we will have
multiple schedulers emitting requests, we can't use that any more.
Remove any remaining uses of it.

The NodeRequest uid was only used to index that dictionary (and
was used to persist a request across resubmission).  Since it isn't
needed any more, it is removed.

Change-Id: I7c82485d95979c6c9a246c3dc3954bae3c65ac13