Fix unstable cross-scheduler config priming test
Since we currently perform a smart reconfiguration after config priming
the scheduler will also update the system config.

This lead to a race condition in
`TestScaleOutScheduler.test_config_priming` as it didn't anticipate that
the first scheduler needs to update its system config again.

To simplify further test cases with multiple schedulers,
`waitUntilSettled()` will now also wait for all schedulers to be primed
before settling.

This change also corrects the `is_valid` property of the system config
cache to check for ltime > -1 (instead of 0). That bug did not have any
consequences in practice.

Change-Id: I2d9acbe8e6002d1c6f6d43b5ab560b533a490e1f